microservices/chat-backend/requirements.txt
asyncpg==0.29.0
confluent_kafka==2.6.0
openai==1.51.2
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-asyncpg==0.48.b